chore: enable i18n for storybook
